Rename Pollard-Rho factorization.

+++ Decompose.Numerics.Test/NumericsTests.cs
@@ -89,8 +89,8 @@ namespace Decompose.Numerics.Test
             TestPrimality3<UInt128>(UInt128.MaxValue - 10000 - 1, 10000, primalityOld, new UInt128MontgomeryReduction());
             TestPrimality3<BigInteger>(0, 10000, primalityOld, new BigIntegerReduction());
             TestPrimality3<BigInteger>(0, 10000, primalityOld, new BigIntegerMontgomeryReduction());
-            TestPrimality3<BigInteger>((BigInteger)ulong.MaxValue * ulong.MaxValue, 10000, primalityOld, new BigIntegerReduction());
-            TestPrimality3<BigInteger>((BigInteger)ulong.MaxValue * ulong.MaxValue, 10000, primalityOld, new BigIntegerMontgomeryReduction());
+            TestPrimality3<BigInteger>((BigInteger)UInt128.MaxValue * UInt128.MaxValue, 10000, primalityOld, new BigIntegerReduction());
+            TestPrimality3<BigInteger>((BigInteger)UInt128.MaxValue * UInt128.MaxValue, 10000, primalityOld, new BigIntegerMontgomeryReduction());
         }
 
         private void TestPrimality3<T>(T startValue, T countValue, IPrimalityAlgorithm<T> primalityOld, IReductionAlgorithm<T> reduction)
